+++ legends/saveslot.py
@@ -1,4 +1,4 @@
-"""The SaveSlot class and supporting objects.
+"""The `legends.saveslot.SaveSlot` class and supporting objects.
 
 """
 
@@ -26,8 +26,8 @@ def ticksToTimedelta(ticks):
     return timedelta(microseconds=ticks//10)
 
 def ticksToDatetime(ticks):
-    """Converts a .NET timestamp to a Python `datetime` object. .NET
-    timestamps return the number of "ticks" since 1/1/0001. There are 10
+    """Converts a .NET timestamp to a Python `datetime` object. A .NET
+    timestamp returns the number of "ticks" since 1/1/0001. There are 10
     "ticks" in a microsecond. (For comparison, a POSIX timestamp returns
     the number of seconds since 1/1/1970.)
 
@@ -44,45 +44,35 @@ def ticksToDatetime(ticks):
     )
 
 class SaveSlot():
-    """One of three player save slots.
+    """Data from one of three save slots in an STL save file.
 
     Attributes:
         timestamps (STLTimeStamps): Stores the timestamp data for the
             save slot.
-        roster (Roster): A Roster object built from the save slot data.
-        tokens (dict of str:int0: A dictionary mapping a character
-            nameID to the number of tokens for that character in the
-            player's possession.
-        favorites (list of Character): A list of characters the player
-            has marked as 'favorite'.
+        roster (legends.roster.Roster): A Roster object built from the
+            save slot data.
+        tokens (dict): {`str`:`int`} A dictionary mapping a character's
+            name ID to the number of tokens for that character in the
+            player's inventory.
+        favorites (list of legends.gameobjects.Character): A list of
+            characters the player has marked as 'favorite'.
 
     """
 
     def __init__(self):
-        """Constructs a SaveSlot object by extracting data from the
-        user's Star Trek: Legends save file, stored on the local HD.
-
-        Args:
-            slot (int): The 0-based index of the save slot from which to
-                draw the data.
-
-        Raises:
-            ValueError: If no save data is found in the given slot.
-
-        """
         self.timestamps = STLTimeStamps()
         self.roster = Roster()
         self.tokens = {nameID: 0 for nameID in GSCharacter}
         self.favorites = []
 
     def fromFile(self, save, slot):
-        """Populates the calling instance's attribute with data from the
-        locally stored save file.
+        """Uses the given save data to population the calling instance's
+        attributes.
 
         Args:
             save (dict): A decrypted dictionary representation of the
-                player's save file, as returned by the `decryptSaveFile`
-                function.
+                player's save file, as returned by the
+                `legends.functions.decryptSaveFile` function.
             slot (int): The 0-based index of the save slot from which to
                 draw the data.
 
@@ -96,35 +86,35 @@ class SaveSlot():
             self.tokens[nameID] = save[key]['items'].get(nameID, 0)
 
     def sort(self, func, descending=True):
-        """Sorts the dictionary of characters stored in the associated
-        Roster object according the currently selected sorting field.
+        """Sorts the dictionary of characters stored in the `roster`
+        attribute according the given function.
 
         Args:
-            func (function): Should be a function mapping a Character
-                object and a SaveSlot object to a sortable value.
+            func (function): Should be a function that maps a
+                `legends.gameobjects.Character` to a sortable value.
+            descending (bool): True if the characters are to be sorted
+                in descending order.
 
         """
         self.roster.chars = dict(sorted(
             self.roster.chars.items(),
-            key=lambda item:func(item[1], self),
+            key=lambda item:func(item[1]),
             reverse=descending
         ))
-        # rewrite sort method in rostertab
-
 
 class STLTimeStamps():
     """An object for storing and managing Star Trek: Legends timestamps.
 
-    An STLTimeStamps object is associated to a specific save slot in a
-    particular user's save file.
+    Each instance is associated to a specific save slot in a particular
+    user's save file.
 
     Attributes:
         startDate (datetime): The time the user first played the
             associated save slot. Defaults to launch of Star Trek:
             Legends.
         timeLastPlayed (datetime): The time the user last played the
-            associated save slot. Defaults to the time the STLTimeStamps
-            instance is created.
+            associated save slot. Defaults to the time the instance is
+            created.
         playDuration (timedelta): The amount of time the user has spent
             playing the associated save slot. Defaults to 0.
 
@@ -137,8 +127,8 @@ class STLTimeStamps():
 
     @property
     def playTimePerDay(self):
-        """timedelta: The amount of time per day spent on the associated
-        save slot.
+        """`timedelta`: The amount of time per day spent on the
+        associated save slot.
         """
         return (
             self.playDuration/(self.timeLastPlayed - self.startDate).days
@@ -146,12 +136,12 @@ class STLTimeStamps():
 
     def fromSaveData(self, save, slot):
         """Sets the attributes of the calling instance to match the data
-        contained in the given save slot of the give save file data.
+        contained in the given save data.
 
         Args:
             save (dict): A decrypted dictionary representation of the
-                player's save file, as returned by the `decryptSaveFile`
-                function.
+                player's save file, as returned by the
+                `legends.functions.decryptSaveFile` function.
             slot (int): The 0-based index of the save slot from which to
                 read the time stamps.